page redesign implementation at 80% * in the backend, the rule that formats the token chain was moved to its own file. * some unused code was removed in timestamp-parser.js
