+++ client/components/editor/editor-modal-editorselect.vue
@@ -1,73 +1,63 @@
 <template lang='pug'>
-  v-dialog(v-model='isShown', persistent, max-width='700', no-click-animation)
-    v-btn(fab, fixed, bottom, right, color='grey darken-3', dark, @click='goBack', style='width: 50px;'): v-icon mdi-undo-variant
-    v-card.radius-7(color='blue darken-3', dark)
-      v-card-text.text-center.py-4
-        .subtitle-1.white--text {{$t('editor:select.title')}}
-        v-container(grid-list-lg, fluid)
-          v-layout(row, wrap, justify-center)
-            //- v-flex(xs4)
-            //-   v-card.radius-7.animated.fadeInUp.wait-p1s(
-            //-     hover
-            //-     light
-            //-     ripple
-            //-     )
-            //-     v-card-text.text-center(@click='selectEditor("video")')
-            //-       img(src='/_assets/svg/editor-icon-markdown.svg', alt='Video', style='width: 36px;')
-            //-       .body-2.primary--text.mt-2 Video
-            //-       .caption.grey--text Transcribe a video
-            v-flex(xs4)
-              v-card.radius-7.animated.fadeInUp.wait-p1s(
-                hover
-                light
-                ripple
-                )
-                v-card-text.text-center(@click='selectEditor("markdown")')
-                  img(src='/_assets/svg/editor-icon-markdown.svg', alt='Markdown', style='width: 36px;')
-                  .body-2.primary--text.mt-2 Markdown
-                  .caption.grey--text Plain Text Formatting
-            v-flex(xs4)
-              v-card.radius-7.animated.fadeInUp.wait-p2s(
-                hover
-                light
-                ripple
-                )
-                v-card-text.text-center(@click='selectEditor("ckeditor")')
-                  img(src='/_assets/svg/editor-icon-ckeditor.svg', alt='Visual Editor', style='width: 36px;')
-                  .body-2.mt-2.primary--text Visual Editor
-                  .caption.grey--text Rich-text WYSIWYG
-            v-flex(xs4)
-              v-card.radius-7.animated.fadeInUp.wait-p3s(
-                hover
-                light
-                ripple
-                )
-                v-card-text.text-center(@click='selectEditor("asciidoc")')
-                  img(src='/_assets/svg/editor-icon-asciidoc.svg', alt='AsciiDoc', style='width: 36px;')
-                  .body-2.primary--text.mt-2 AsciiDoc
-                  .caption.grey--text Plain Text Formatting
-            v-flex(xs4)
-              v-card.radius-7.animated.fadeInUp.wait-p4s(
-                hover
-                light
-                ripple
-                )
-                v-card-text.text-center(@click='selectEditor("code")')
-                  img(src='/_assets/svg/editor-icon-code.svg', alt='Code', style='width: 36px;')
-                  .body-2.primary--text.mt-2 Code
-                  .caption.grey--text Raw HTML
-            v-flex(xs4)
-              v-card.radius-7.animated.fadeInUp.wait-p5s(
-                hover
-                light
-                ripple
-                )
-                v-card-text.text-center(@click='fromTemplate')
-                  img(src='/_assets/svg/icon-cube.svg', alt='From Template', style='width: 42px; opacity: .5;')
-                  .body-2.mt-1.teal--text From Template
-                  .caption.grey--text Use an existing page...
-
-    page-selector(mode='select', v-model='templateDialogIsShown', :open-handler='fromTemplateHandle', :path='path', :locale='locale', must-exist)
+  <!-- v-dialog(v-model='isShown', persistent, max-width='700', no-click-animation) -->
+  <!--   v-btn(fab, fixed, bottom, right, color='grey darken-3', dark, @click='goBack', style='width: 50px;'): v-icon mdi-undo-variant -->
+  <!--   v-card.radius-7(color='blue darken-3', dark) -->
+  <!--     v-card-text.text-center.py-4 -->
+  <!--       .subtitle-1.white--text {{$t('editor:select.title')}} -->
+  <!--       v-container(grid-list-lg, fluid) -->
+  <!--         v-layout(row, wrap, justify-center) -->
+  <!--           v-flex(xs6) -->
+  <!--             v-card.radius-7.animated.fadeInUp.wait-p1s( -->
+  <!--               hover -->
+  <!--               light -->
+  <!--               ripple -->
+  <!--               ) -->
+  <!--               v-card-text.text-center(@click='selectEditor("markdown")') -->
+  <!--                 img(src='/_assets/svg/editor-icon-markdown.svg', alt='Markdown', style='width: 36px;') -->
+  <!--                 .body-2.primary--text.mt-2 Markdown -->
+  <!--                 .caption.grey--text Plain Text Formatting -->
+  <!--           v-flex(xs6) -->
+  <!--             v-card.radius-7.animated.fadeInUp.wait-p2s( -->
+  <!--               hover -->
+  <!--               light -->
+  <!--               ripple -->
+  <!--               ) -->
+  <!--               v-card-text.text-center(@click='selectEditor("ckeditor")') -->
+  <!--                 img(src='/_assets/svg/editor-icon-ckeditor.svg', alt='Visual Editor', style='width: 36px;') -->
+  <!--                 .body-2.mt-2.primary--text Visual Editor -->
+  <!--                 .caption.grey--text Rich-text WYSIWYG -->
+  <!--           v-flex(xs4) -->
+  <!--             v-card.radius-7.animated.fadeInUp.wait-p3s( -->
+  <!--               hover -->
+  <!--               light -->
+  <!--               ripple -->
+  <!--               ) -->
+  <!--               v-card-text.text-center(@click='selectEditor("asciidoc")') -->
+  <!--                 img(src='/_assets/svg/editor-icon-asciidoc.svg', alt='AsciiDoc', style='width: 36px;') -->
+  <!--                 .body-2.primary--text.mt-2 AsciiDoc -->
+  <!--                 .caption.grey--text Plain Text Formatting -->
+  <!--           v-flex(xs4) -->
+  <!--             v-card.radius-7.animated.fadeInUp.wait-p4s( -->
+  <!--               hover -->
+  <!--               light -->
+  <!--               ripple -->
+  <!--               ) -->
+  <!--               v-card-text.text-center(@click='selectEditor("code")') -->
+  <!--                 img(src='/_assets/svg/editor-icon-code.svg', alt='Code', style='width: 36px;') -->
+  <!--                 .body-2.primary--text.mt-2 Code -->
+  <!--                 .caption.grey--text Raw HTML -->
+  <!--           v-flex(xs4) -->
+  <!--             v-card.radius-7.animated.fadeInUp.wait-p5s( -->
+  <!--               hover -->
+  <!--               light -->
+  <!--               ripple -->
+  <!--               ) -->
+  <!--               v-card-text.text-center(@click='fromTemplate') -->
+  <!--                 img(src='/_assets/svg/icon-cube.svg', alt='From Template', style='width: 42px; opacity: .5;') -->
+  <!--                 .body-2.mt-1.teal--text From Template -->
+  <!--                 .caption.grey--text Use an existing page... -->
+  <!---->
+  <!--   page-selector(mode='select', v-model='templateDialogIsShown', :open-handler='fromTemplateHandle', :path='path', :locale='locale', must-exist) -->
 </template>
 
 <script>
@@ -97,10 +87,7 @@ export default {
   },
   methods: {
     selectEditor (name) {
-      const currentLocation = new URLSearchParams(window.location.search);
-      const params = Object.fromEntries(currentLocation.entries());
-      // console.log(params);
-      this.currentEditor = `editor${_.startCase(params.editor)}`
+      this.currentEditor = `editor${_.startCase(name)}`
       this.isShown = false
     },
     goBack () {
@@ -120,8 +107,7 @@ export default {
   mounted () {
     const currentLocation = new URLSearchParams(window.location.search);
     const params = Object.fromEntries(currentLocation.entries());
-    // console.log(params);
-    this.currentEditor = `editor${_.startCase(params.editor)}`
+    if (Object.keys(params).length) this.currentEditor = `editor${_.startCase(params.editor)}`
   }
 }
 </script>
